remove replace winio old version & separate hns call into windows-only file
