+++ cns/service/main.go
@@ -63,7 +63,6 @@ import (
 	localtls "github.com/Azure/azure-container-networking/server/tls"
 	"github.com/Azure/azure-container-networking/store"
 	"github.com/Azure/azure-container-networking/telemetry"
-	"github.com/Microsoft/hcsshim"
 	"github.com/avast/retry-go/v4"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
@@ -493,10 +492,6 @@ func startTelemetryService(ctx context.Context) {
 	tb.PushData(rootCtx)
 }
 
-type wscliInterface interface {
-	GetInterfaces(ctx context.Context) (*wireserver.GetInterfacesResult, error)
-}
-
 // Main is the entry point for CNS.
 func main() {
 	// Initialize and parse command line arguments.
@@ -882,16 +877,9 @@ func main() {
 			}
 
 			// Setting primary macaddress if VF is enabled on the nics for swiftv2
-			var wscli wscliInterface
-			// SWIFT V2 mode for accelnet, supply the MAC address to the HNS
-			macAddress, err := getPrimaryNICMACAddress(wscli)
+			err = setVFForAccelnetNICs()	
 			if err != nil {
-				logger.Errorf("Failed to get primary NIC MAC address: %v", err)
-			} else {
-				macAddresses := []string{macAddress}
-				if _, err := hcsshim.SetNnvManagementMacAddresses(macAddresses); err != nil {
-					logger.Errorf("Failed to set primary NIC MAC address: %v", err)
-				}
+				logger.Errorf("Failed to get primary NIC MAC address & set VF for accelnet NICs: %v", err)
 			}
 		}
 	}
@@ -1631,29 +1619,4 @@ func PopulateCNSEndpointState(endpointStateStore store.KeyValueStore) error {
 		return fmt.Errorf("failed to write endpoint state to store: %w", err)
 	}
 	return nil
-}
-
-// getPrimaryNICMacAddress fetches the MAC address of the primary NIC on the node.
-func getPrimaryNICMACAddress(wscli wscliInterface) (string, error) {
-	res, err := wscli.GetInterfaces(context.TODO())
-	if err != nil {
-		return "", fmt.Errorf("failed to find primary interface info: %w", err)
-	}
-	var macAddress string
-	for _, i := range res.Interface {
-		// skip if not primary
-		if !i.IsPrimary {
-			continue
-		}
-		// skip if no subnets
-		if len(i.IPSubnet) == 0 {
-			continue
-		}
-		macAddress = i.MacAddress
-	}
-
-	if macAddress == "" {
-		return "", errors.New("MAC address not found in wscli")
-	}
-	return macAddress, nil
 }
