+++ cns/service/main_linux.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"errors"
+)
+
+func setVFForAccelnetNICs() error {
+	return nil
+}
